fix(systemMemorySizeMb): Improve system memory detection on BSD systems
- Added support for detecting total system memory on FreeBSD, OpenBSD, and NetBSD using `sysctl`.
- Updated `systemMemorySizeMb()` to use `unsigned long` instead of `int` for better accuracy, especially on systems with large amounts of RAM.
- Introduced a `constexpr unsigned long megabyte` to improve readability and avoid redundant calculations.
- Unified memory detection logic for BSD systems, using `HW_PHYSMEM64` when available for better precision.

Fixes incorrect memory reporting on BSD platforms where it previously returned `-1`.

+++ src/core/qgsimagecache.cpp
@@ -102,7 +102,7 @@ QgsImageCache::QgsImageCache( QObject *parent )
   }
   else
   {
-    const int sysMemory = QgsApplication::systemMemorySizeMb();
+    const unsigned long sysMemory = QgsApplication::systemMemorySizeMb();
     if ( sysMemory > 0 )
     {
       if ( sysMemory >= 32000 ) // 32 gb RAM (or more) = 500mb cache size
